auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
PyJWT==2.8.0
redis==5.0.1
python-dotenv==1.0.0
alembic==1.13.1
httpx==0.25.0
authlib==1.2.1
email-validator==2.1.0